Error is in the methodology of the bus.
In axi slave implementation the clock speed is very high.
The master will never return to the idle state after going to s2 or s3, since enable will be high for a large number of clock cycles.
Need to redo the design such that the idle => await only happens with posedge of enable.
Master should return to idle after every read or write if the enable is low.
 Changes to be committed:
	modified:   ahb.srcs/sim_1/imports/AHB/wrapper_tb.v
	modified:   ahb.srcs/sources_1/bd/design_1/hdl/design_1_wrapper.v
	modified:   ahb.srcs/sources_1/imports/AHB/ahb_master.v
	renamed:    deploy_bus/deploy/blockdesign.tcl -> deploy_bus/design_1.tcl
	modified:   deploy_bus/ip_repo/amba_ahb_1.0/component.xml
	modified:   deploy_bus/ip_repo/amba_ahb_1.0/hdl/amba_ahb_v1_0.v
	modified:   deploy_bus/ip_repo/amba_ahb_1.0/hdl/amba_ahb_v1_0_S_AXI.v
	modified:   deploy_bus/ip_repo/amba_ahb_1.0/src/ahb_master.v
	modified:   deploy_bus/ip_repo/amba_ahb_1.0/src/design_1_2.v
	modified:   deploy_bus/ip_repo/amba_ahb_1.0/user.org_user_amba_ahb_5.0.zip
	new file:   deploy_bus/ip_repo/amba_ahb_1.0/user.org_user_amba_ahb_6.0.zip
	new file:   deploy_bus/ip_repo/amba_ahb_1.0/xgui/amba_ahb_v6_0.tcl